Update RtD and Sphinx configuration  (#38046)
